\name{print.eivlm}
\alias{print.eivlm}

\title{\code{print} method for objects of class \code{eivlm}.}

\description{
  Analogous to \code{print} method for \code{\link{lm}} objects.
}
  
\usage{
\method{print}{eivlm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}

\arguments{
  \item{x}{
    A model object of class \code{eivlm}.
  }
  
  \item{digits, \dots}{
    See help for \code{\link{print.default}}.
  }
}

\keyword{methods}

