\name{vcov.eivlm}
\alias{vcov.eivlm}

\title{\code{vcov} method for objects of class \code{eivlm}.}

\description{
  Extract variance/covariance matrix of estimated parameters from
  \code{eivlm} model object. Analogous to \code{vcov} method for other models.
}

\usage{
\method{vcov}{eivlm}(object, ...)
}

\arguments{
  \item{object}{
    A model object of class \code{eivlm}.
  }
  
  \item{\dots}{
    Not currently implemented.
  }
}

\value{
  Estimated variance/covariance matrix of estimated regression
  coefficients.
}

\keyword{methods}


