\name{eivtools-package}
\alias{eivtools-package}
\alias{eivtools}
\docType{package}

\title{
  \packageTitle{eivtools}
}

\description{
  \packageDescription{eivtools}
}

\details{
  %% The DESCRIPTION file:
  %% \packageDESCRIPTION{eivtools}
  %% \packageIndices{eivtools}
  %% ~~ An overview of how to use the package, including the most important ~~
  %% ~~ functions ~~
  Forthcoming.
}

\author{
  \packageAuthor{eivtools}
  
  Maintainer: \packageMaintainer{eivtools}
}

\references{
  R.J. Carroll, D. Ruppert, L.A. Stefanski and C.M. Crainiceanu
  (2006). \emph{Measurement Error in Nonlinear Models: A Modern
    Perspective} (2nd edition).  London: Chapman & Hall.
  
  W. Fuller (2006). \emph{Measurement Error Models} (2nd edition). New York: John
  Wiley & Sons.
  
  L.A. Stefanksi and D.B. Boos (2002). \dQuote{The calculus of
    {M}-estimation,} \emph{The American Statistician} 56(1):29-38.

  J. Wooldridge (2002). \emph{Econometric Analysis of Cross Section and
  Panel Data}. Cambridge, MA: MIT Press.
}

\keyword{ package }

%% \seealso{
  %% ~~ Optional links to other man pages, e.g. ~~
  %% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }

%% \examples{
  %% ~~ simple examples of the most important functions ~~
%% }
