\name{eivtools-package}
\alias{eivtools-package}
\alias{eivtools}
\docType{package}
\title{
  \packageTitle{eivtools}
}
\description{
  \packageDescription{eivtools}
}
\details{
  %% The DESCRIPTION file:
  %% \packageDESCRIPTION{eivtools}
  %% \packageIndices{eivtools}
  %% ~~ An overview of how to use the package, including the most important ~~
  %% ~~ functions ~~
  TBD
}
\author{
  \packageAuthor{eivtools} 
  Maintainer: \packageMaintainer{eivtools}
}
\references{
  %% ~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
  %% ~~ Optional links to other man pages, e.g. ~~
  %% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
  %% ~~ simple examples of the most important functions ~~
}
