\name{eivtools-package}
\alias{eivtools-package}
\alias{eivtools}
\docType{package}

\title{
  \packageTitle{eivtools}
}

\description{
  \packageDescription{eivtools}
}

\details{
  %% The DESCRIPTION file:
  %% \packageDESCRIPTION{eivtools}
  %% \packageIndices{eivtools}
  %% ~~ An overview of how to use the package, including the most important ~~
  %% ~~ functions ~~
  Forthcoming.
}

\author{
  \packageAuthor{eivtools}
  
  Maintainer: \packageMaintainer{eivtools}
}

\references{
  Carroll R.J, Ruppert D., Stefanski L.A. and Crainiceanu C.M.
  (2006). \emph{Measurement Error in Nonlinear Models: A Modern
    Perspective} (2nd edition).  London: Chapman & Hall.
  
  Fuller W. (2006). \emph{Measurement Error Models} (2nd edition). New York: John
  Wiley & Sons.

  Lockwood J.R and McCaffrey D.F. (2014). \dQuote{Correcting for test
    score measurement error in ANCOVA models for estimating treatment
    effects,} \emph{Journal of Educational and Behavioral Statistics}
  39(1):22-52.

  Lockwood J.R. and McCaffrey D.F. (forthcoming).
  \dQuote{Simulation-extrapolation with latent heteroskedastic
    variance,} \emph{Psychometrika}.
  
  Rabe-Hesketh S., Pickles A. and Skrondal A. (2003). \dQuote{Correcting
  for covariate measurement error in logistic regression using
  nonparametric maximum likelihood estimation,} \emph{Statistical
  Modelling} 3:215-232.

  Stefanksi L.A. and Boos D.B. (2002). \dQuote{The calculus of
    {M}-estimation,} \emph{The American Statistician} 56(1):29-38.

  Wooldridge J. (2002). \emph{Econometric Analysis of Cross Section and
  Panel Data}. Cambridge, MA: MIT Press.
}

\keyword{ package }

%% \seealso{
  %% ~~ Optional links to other man pages, e.g. ~~
  %% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }

%% \examples{
  %% ~~ simple examples of the most important functions ~~
%% }
